Refaktorering etter kommentar på PR Fiks av test på tidspunkt
