+++ src/main/kotlin/no/nav/familie/oppdrag/service/KonsistensavstemmingService.kt
@@ -3,13 +3,12 @@ package no.nav.familie.oppdrag.service
 import no.nav.familie.kontrakter.felles.oppdrag.Utbetalingsoppdrag
 import no.nav.familie.oppdrag.avstemming.AvstemmingSender
 import no.nav.familie.oppdrag.konsistensavstemming.KonsistensavstemmingMapper
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 import java.time.LocalDateTime
 
 @Service
 class KonsistensavstemmingService(
-        @Autowired private val avstemmingSender: AvstemmingSender) {
+        private val avstemmingSender: AvstemmingSender) {
 
     fun utførKonsistensavstemming(fagsystem: String, utbetalingsoppdrag: List<Utbetalingsoppdrag>, avstemmingsdato: LocalDateTime) {
         val konsistensavstemmingMapper = KonsistensavstemmingMapper(fagsystem, utbetalingsoppdrag, avstemmingsdato)
